+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v1.0.1 - 2025-01-13
+
+Internal refactoring.
+
+This fixes a theoretically possible but statistically unlikely bug where `randline` could return less than *k* items, even when there were less than *k* items in the input.
+
 ## v1.0.0 - 2025-01-11
 
 Initial release.